+++ src/SortableContainer/index.js
@@ -9,7 +9,10 @@ import {
   events,
   vendorPrefix,
   limit,
+  getEdgeOffset,
   getElementMargin,
+  getLockPixelOffset,
+  getPosition,
   provideDisplayName,
   omit,
 } from '../utils';
@@ -100,7 +103,6 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
 
     componentDidMount() {
       const {
-        getContainer,
         useWindowAsScrollContainer,
       } = this.props;
 
@@ -110,27 +112,29 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
        *  https://github.com/clauderic/react-sortable-hoc/issues/249
        */
 
-      this.container = typeof getContainer === 'function'
-        ? getContainer(this.getWrappedInstance())
-        : findDOMNode(this);
-      this.document = this.container.ownerDocument || document;
+      const container = this.getContainer();
 
-      const contentWindow = this.props.contentWindow || this.document.defaultView || window;
+      Promise.resolve(container).then((containerNode) => {
+        this.container = containerNode;
+        this.document = this.container.ownerDocument || document;
 
-      this.contentWindow = typeof contentWindow === 'function'
-        ? contentWindow()
-        : contentWindow;
-      this.scrollContainer = useWindowAsScrollContainer
-        ? this.document.scrollingElement || this.document.documentElement
-        : this.container;
+        const contentWindow = this.props.contentWindow || this.document.defaultView || window;
 
-      for (const key in this.events) {
-        if (this.events.hasOwnProperty(key)) {
-          events[key].forEach(eventName =>
-            this.container.addEventListener(eventName, this.events[key], false)
-          );
+        this.contentWindow = typeof contentWindow === 'function'
+          ? contentWindow()
+          : contentWindow;
+        this.scrollContainer = useWindowAsScrollContainer
+          ? this.document.scrollingElement || this.document.documentElement
+          : this.container;
+
+        for (const key in this.events) {
+          if (this.events.hasOwnProperty(key)) {
+            events[key].forEach(eventName =>
+              this.container.addEventListener(eventName, this.events[key], false)
+            );
+          }
         }
-      }
+      });
     }
 
     componentWillUnmount() {
@@ -152,20 +156,17 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
       }
     }
 
-    handleStart = e => {
+    handleStart = event => {
       const {distance, shouldCancelStart} = this.props;
 
-      if (e.button === 2 || shouldCancelStart(e)) {
+      if (event.button === 2 || shouldCancelStart(event)) {
         return false;
       }
 
       this._touched = true;
-      this._pos = {
-        x: e.pageX,
-        y: e.pageY,
-      };
+      this._pos = getPosition(event);
 
-      const node = closest(e.target, el => el.sortableInfo != null);
+      const node = closest(event.target, el => el.sortableInfo != null);
 
       if (
         node &&
@@ -177,27 +178,27 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
         const {index, collection} = node.sortableInfo;
 
         if (
-          useDragHandle && !closest(e.target, el => el.sortableHandle != null)
+          useDragHandle && !closest(event.target, el => el.sortableHandle != null)
         )
           return;
 
         this.manager.active = {index, collection};
 
         /*
-         * Fixes a bug in Firefox where the :active state of anchor tags
-         * prevent subsequent 'mousemove' events from being fired
-         * (see https://github.com/clauderic/react-sortable-hoc/issues/118)
-         */
-        if (e.target.tagName.toLowerCase() === 'a') {
-          e.preventDefault();
+				 * Fixes a bug in Firefox where the :active state of anchor tags
+				 * prevent subsequent 'mousemove' events from being fired
+				 * (see https://github.com/clauderic/react-sortable-hoc/issues/118)
+				 */
+        if (event.target.tagName.toLowerCase() === 'a') {
+          event.preventDefault();
         }
 
         if (!distance) {
           if (this.props.pressDelay === 0) {
-            this.handlePress(e);
+            this.handlePress(event);
           } else {
             this.pressTimer = setTimeout(
-              () => this.handlePress(e),
+              () => this.handlePress(event),
               this.props.pressDelay
             );
           }
@@ -209,21 +210,22 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
       return node.sortableInfo.manager === this.manager;
     };
 
-    handleMove = e => {
+    handleMove = event => {
       const {distance, pressThreshold} = this.props;
 
       if (!this.state.sorting && this._touched) {
-        this._delta = {
-          x: this._pos.x - e.pageX,
-          y: this._pos.y - e.pageY,
+        const position = getPosition(event);
+        const delta = this._delta = {
+          x: this._pos.x - position.x,
+          y: this._pos.y - position.y,
         };
-        const delta = Math.abs(this._delta.x) + Math.abs(this._delta.y);
+        const combinedDelta = Math.abs(delta.x) + Math.abs(delta.y);
 
-        if (!distance && (!pressThreshold || pressThreshold && delta >= pressThreshold)) {
+        if (!distance && (!pressThreshold || pressThreshold && combinedDelta >= pressThreshold)) {
           clearTimeout(this.cancelTimer);
           this.cancelTimer = setTimeout(this.cancel, 0);
-        } else if (distance && delta >= distance && this.manager.isActive()) {
-          this.handlePress(e);
+        } else if (distance && combinedDelta >= distance && this.manager.isActive()) {
+          this.handlePress(event);
         }
       }
     };
@@ -245,7 +247,7 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
       }
     };
 
-    handlePress = e => {
+    handlePress = event => {
       const active = this.manager.getActive();
 
       if (active) {
@@ -281,8 +283,8 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
           x: axis.indexOf('x') >= 0,
           y: axis.indexOf('y') >= 0,
         };
-        this.offsetEdge = this.getEdgeOffset(node);
-        this.initialOffset = this.getOffset(e);
+        this.offsetEdge = getEdgeOffset(node, this.container);
+        this.initialOffset = getPosition(event);
         this.initialScroll = {
           top: this.container.scrollTop,
           left: this.container.scrollLeft,
@@ -352,7 +354,7 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
           this.helper.classList.add(...helperClass.split(' '));
         }
 
-        this.listenerNode = e.touches ? node : this.contentWindow;
+        this.listenerNode = event.touches ? node : this.contentWindow;
         events.move.forEach(eventName =>
           this.listenerNode.addEventListener(
             eventName,
@@ -371,23 +373,27 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
           sortingIndex: index,
         });
 
-        if (onSortStart) onSortStart({node, index, collection}, e);
+        if (onSortStart) {
+          onSortStart({node, index, collection}, event);
+        }
       }
     };
 
-    handleSortMove = e => {
+    handleSortMove = event => {
       const {onSortMove} = this.props;
-      e.preventDefault(); // Prevent scrolling on mobile
+      event.preventDefault(); // Prevent scrolling on mobile
 
-      this.updatePosition(e);
+      this.updatePosition(event);
       this.animateNodes();
       this.autoscroll();
 
-      if (onSortMove) onSortMove(e);
+      if (onSortMove) {
+        onSortMove(event);
+      }
     };
 
-    handleSortEnd = e => {
-      const {transitionDuration, onSortEnd} = this.props;
+    handleSortEnd = event => {
+      const {hideSortableGhost, transitionDuration, onSortEnd} = this.props;
       const {collection} = this.manager.active;
 
       // Remove the event listeners if the node is still in the DOM
@@ -444,7 +450,7 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
             newIndex: this.newIndex,
             collection,
           },
-          e
+          event
         );
       }
 
@@ -476,106 +482,38 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
       // Update manager state
       this.manager.active = null;
     };
-    getEdgeOffset(node, offset = {top: 0, left: 0}) {
-      // Get the actual offsetTop / offsetLeft value, no matter how deep the node is nested
-      if (node) {
-        const nodeOffset = {
-          top: offset.top + node.offsetTop,
-          left: offset.left + node.offsetLeft,
-        };
-        if (node.parentNode !== this.container) {
-          return this.getEdgeOffset(node.parentNode, nodeOffset);
-        } else {
-          return nodeOffset;
-        }
-      }
-    }
-
-    getOffset(e) {
-      if (e.touches && e.touches.length) {
-        return {
-          x: e.touches[0].pageX,
-          y: e.touches[0].pageY,
-        };
-      } else if (e.changedTouches && e.changedTouches.length) {
-        return {
-          x: e.changedTouches[0].pageX,
-          y: e.changedTouches[0].pageY,
-        };
-      } else {
-        return {
-          x: e.pageX,
-          y: e.pageY,
-        };
-      }
-    }
 
     getLockPixelOffsets() {
-      let {lockOffset} = this.props;
-
-      if (!Array.isArray(lockOffset)) {
-        lockOffset = [lockOffset, lockOffset];
-      }
+      const {width, height} = this;
+      const {lockOffset} = this.props;
+      const offsets = Array.isArray(lockOffset)
+        ? lockOffset
+        : [lockOffset, lockOffset];
 
       invariant(
-        lockOffset.length === 2,
+        offsets.length === 2,
         'lockOffset prop of SortableContainer should be a single ' +
           'value or an array of exactly two values. Given %s',
         lockOffset
       );
 
-      const [minLockOffset, maxLockOffset] = lockOffset;
+      const [minLockOffset, maxLockOffset] = offsets;
 
       return [
-        this.getLockPixelOffset(minLockOffset),
-        this.getLockPixelOffset(maxLockOffset),
+        getLockPixelOffset({lockOffset: minLockOffset, width, height}),
+        getLockPixelOffset({lockOffset: maxLockOffset, width, height}),
       ];
     }
 
-    getLockPixelOffset(lockOffset) {
-      let offsetX = lockOffset;
-      let offsetY = lockOffset;
-      let unit = 'px';
-
-      if (typeof lockOffset === 'string') {
-        const match = /^[+-]?\d*(?:\.\d*)?(px|%)$/.exec(lockOffset);
-
-        invariant(
-          match !== null,
-          'lockOffset value should be a number or a string of a ' +
-            'number followed by "px" or "%". Given %s',
-          lockOffset
-        );
-
-        offsetX = (offsetY = parseFloat(lockOffset));
-        unit = match[1];
-      }
-
-      invariant(
-        isFinite(offsetX) && isFinite(offsetY),
-        'lockOffset value should be a finite. Given %s',
-        lockOffset
-      );
-
-      if (unit === '%') {
-        offsetX = offsetX * this.width / 100;
-        offsetY = offsetY * this.height / 100;
-      }
-
-      return {
-        x: offsetX,
-        y: offsetY,
-      };
-    }
-
-    updatePosition(e) {
+    updatePosition(event) {
       const {lockAxis, lockToContainerEdges} = this.props;
 
-      const offset = this.getOffset(e);
+      const offset = getPosition(event);
       const translate = {
         x: offset.x - this.initialOffset.x,
         y: offset.y - this.initialOffset.y,
       };
+
       // Adjust for window scroll
       translate.y -= (window.pageYOffset - this.initialWindowScroll.top);
       translate.x -= (window.pageXOffset - this.initialWindowScroll.left);
@@ -633,7 +571,7 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
       };
 
       const ghostOffset = nodes[this.index].edgeOffset ||
-                (nodes[this.index].edgeOffset = this.getEdgeOffset(nodes[this.index].node));
+                (nodes[this.index].edgeOffset = getEdgeOffset(nodes[this.index].node));
 
       const ghostTranslate = {
         x: 0,
@@ -663,7 +601,7 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
 
         // If we haven't cached the node's offsetTop / offsetLeft value
         if (!edgeOffset) {
-          nodes[i].edgeOffset = (edgeOffset = this.getEdgeOffset(node));
+          nodes[i].edgeOffset = (edgeOffset = getEdgeOffset(node, this.container));
         }
 
         // Get a reference to the next and previous node
@@ -673,7 +611,7 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
         // Also cache the next node's edge offset if needed.
         // We need this for calculating the animation in a grid setup
         if (nextNode && !nextNode.edgeOffset) {
-          nextNode.edgeOffset = this.getEdgeOffset(nextNode.node);
+          nextNode.edgeOffset = getEdgeOffset(nextNode.node, this.container);
         }
 
         if (transitionDuration) {
@@ -871,7 +809,18 @@ export default function sortableContainer(WrappedComponent, config = {withRef: f
         config.withRef,
         'To access the wrapped instance, you need to pass in {withRef: true} as the second argument of the SortableContainer() call'
       );
+
       return this.refs.wrappedInstance;
+    }
+
+    getContainer() {
+      const {getContainer} = this.props;
+
+      if (typeof getContainer !== 'function') {
+        return findDOMNode(this);
+      }
+
+      return getContainer(config.withRef ? this.getWrappedInstance() : undefined);
     }
 
     render() {